moved windows_tts behind cfg(windows)
